refactor hardcoded
