+++ packages/config/src/discovery/hardcoded/test/constants.ts
@@ -1,0 +1,18 @@
+/*
+                      ====== IMPORTANT NOTICE ======
+
+EDIT THIS FILE ONLY WHEN YOU KNOW WHAT YOU ARE DOING
+
+This is a file responsible for hardcoding the data into the config. 
+The data is hardcoded because it is not possible to fetch it from the blockchain, using current discovery methods.
+
+Updating this file should be a conscious decision preceded by the research.
+
+DO NOT UPDATE THIS FILE ONLY TO FIX THE TESTS
+UNDERSTAND WHAT YOU ARE DOING BEFORE YOU UPDATE THIS FILE
+*/
+
+export const TESTS_HARDCODED = {
+  ARBITRUM_SET_VALIDATOR_COUNT: 7,
+  ARBITRUM_SET_SEQUENCER_COUNT: 3,
+}
